+++ src/gui/main_window.cpp
@@ -2,11 +2,21 @@
 /// @brief Main application window implementation
 
 #include "kalahari/gui/main_window.h"
+#include "kalahari/gui/panels/editor_panel.h"
+#include "kalahari/gui/panels/navigator_panel.h"
+#include "kalahari/gui/panels/properties_panel.h"
+#include "kalahari/gui/panels/search_panel.h"
+#include "kalahari/gui/panels/assistant_panel.h"
+#include "kalahari/gui/panels/log_panel.h"
 #include "kalahari/core/logger.h"
 #include <QMenuBar>
 #include <QToolBar>
 #include <QStatusBar>
 #include <QApplication>
+#include <QSettings>
+#include <QDockWidget>
+#include <QCloseEvent>
+#include <QShowEvent>
 
 namespace kalahari {
 namespace gui {
@@ -25,7 +35,26 @@ MainWindow::MainWindow(QWidget* parent)
     , m_pasteAction(nullptr)
     , m_fileMenu(nullptr)
     , m_editMenu(nullptr)
+    , m_viewMenu(nullptr)
     , m_fileToolbar(nullptr)
+    , m_viewNavigatorAction(nullptr)
+    , m_viewPropertiesAction(nullptr)
+    , m_viewLogAction(nullptr)
+    , m_viewSearchAction(nullptr)
+    , m_viewAssistantAction(nullptr)
+    , m_resetLayoutAction(nullptr)
+    , m_navigatorDock(nullptr)
+    , m_propertiesDock(nullptr)
+    , m_logDock(nullptr)
+    , m_searchDock(nullptr)
+    , m_assistantDock(nullptr)
+    , m_editorPanel(nullptr)
+    , m_navigatorPanel(nullptr)
+    , m_propertiesPanel(nullptr)
+    , m_searchPanel(nullptr)
+    , m_assistantPanel(nullptr)
+    , m_logPanel(nullptr)
+    , m_firstShow(true)
 {
     auto& logger = core::Logger::getInstance();
     logger.info("MainWindow constructor called");
@@ -39,6 +68,7 @@ MainWindow::MainWindow(QWidget* parent)
     createMenus();
     createToolbars();
     createStatusBar();
+    createDocks();
 
     logger.info("MainWindow initialized successfully");
 }
@@ -134,11 +164,13 @@ void MainWindow::createToolbars() {
 
     // File toolbar
     m_fileToolbar = addToolBar(tr("File"));
+    m_fileToolbar->setMovable(true);      // Can be moved between dock areas
+    m_fileToolbar->setFloatable(true);    // Can be detached as floating window
     m_fileToolbar->addAction(m_newAction);
     m_fileToolbar->addAction(m_openAction);
     m_fileToolbar->addAction(m_saveAction);
 
-    logger.debug("Toolbars created successfully");
+    logger.debug("Toolbars created successfully (movable and floatable)");
 }
 
 void MainWindow::createStatusBar() {
@@ -209,6 +241,157 @@ void MainWindow::onPaste() {
     auto& logger = core::Logger::getInstance();
     logger.info("Action triggered: Paste");
     statusBar()->showMessage(tr("Paste (not implemented)"), 2000);
+}
+
+// Dock management
+void MainWindow::createDocks() {
+    auto& logger = core::Logger::getInstance();
+    logger.debug("Creating dock widgets");
+
+    // Create central editor panel
+    m_editorPanel = new EditorPanel(this);
+    setCentralWidget(m_editorPanel);
+
+    // Navigator dock (left)
+    m_navigatorPanel = new NavigatorPanel(this);
+    m_navigatorDock = new QDockWidget(tr("Navigator"), this);
+    m_navigatorDock->setWidget(m_navigatorPanel);
+    m_navigatorDock->setObjectName("NavigatorDock");  // Required for saveState!
+    addDockWidget(Qt::LeftDockWidgetArea, m_navigatorDock);
+
+    // Properties dock (right)
+    m_propertiesPanel = new PropertiesPanel(this);
+    m_propertiesDock = new QDockWidget(tr("Properties"), this);
+    m_propertiesDock->setWidget(m_propertiesPanel);
+    m_propertiesDock->setObjectName("PropertiesDock");
+    addDockWidget(Qt::RightDockWidgetArea, m_propertiesDock);
+
+    // Log dock (bottom)
+    m_logPanel = new LogPanel(this);
+    m_logDock = new QDockWidget(tr("Log"), this);
+    m_logDock->setWidget(m_logPanel);
+    m_logDock->setObjectName("LogDock");
+    addDockWidget(Qt::BottomDockWidgetArea, m_logDock);
+
+    // Search dock (right, tabbed with Properties)
+    m_searchPanel = new SearchPanel(this);
+    m_searchDock = new QDockWidget(tr("Search"), this);
+    m_searchDock->setWidget(m_searchPanel);
+    m_searchDock->setObjectName("SearchDock");
+    addDockWidget(Qt::RightDockWidgetArea, m_searchDock);
+    tabifyDockWidget(m_propertiesDock, m_searchDock);
+
+    // Assistant dock (right, tabbed with Properties/Search)
+    m_assistantPanel = new AssistantPanel(this);
+    m_assistantDock = new QDockWidget(tr("Assistant"), this);
+    m_assistantDock->setWidget(m_assistantPanel);
+    m_assistantDock->setObjectName("AssistantDock");
+    addDockWidget(Qt::RightDockWidgetArea, m_assistantDock);
+    tabifyDockWidget(m_searchDock, m_assistantDock);
+
+    // Raise Properties tab (default visible)
+    m_propertiesDock->raise();
+
+    // Create View menu
+    m_viewMenu = menuBar()->addMenu(tr("&View"));
+
+    // Create toggle actions (use QDockWidget's built-in toggleViewAction!)
+    m_viewNavigatorAction = m_navigatorDock->toggleViewAction();
+    m_viewNavigatorAction->setShortcut(QKeySequence(tr("Ctrl+1")));
+    m_viewMenu->addAction(m_viewNavigatorAction);
+
+    m_viewPropertiesAction = m_propertiesDock->toggleViewAction();
+    m_viewPropertiesAction->setShortcut(QKeySequence(tr("Ctrl+2")));
+    m_viewMenu->addAction(m_viewPropertiesAction);
+
+    m_viewLogAction = m_logDock->toggleViewAction();
+    m_viewLogAction->setShortcut(QKeySequence(tr("Ctrl+3")));
+    m_viewMenu->addAction(m_viewLogAction);
+
+    m_viewSearchAction = m_searchDock->toggleViewAction();
+    m_viewSearchAction->setShortcut(QKeySequence(tr("Ctrl+4")));
+    m_viewMenu->addAction(m_viewSearchAction);
+
+    m_viewAssistantAction = m_assistantDock->toggleViewAction();
+    m_viewAssistantAction->setShortcut(QKeySequence(tr("Ctrl+5")));
+    m_viewMenu->addAction(m_viewAssistantAction);
+
+    m_viewMenu->addSeparator();
+
+    // Reset layout action
+    m_resetLayoutAction = new QAction(tr("Reset Layout"), this);
+    m_resetLayoutAction->setShortcut(QKeySequence(tr("Ctrl+0")));
+    m_resetLayoutAction->setStatusTip(tr("Reset dock layout to default"));
+    connect(m_resetLayoutAction, &QAction::triggered, this, &MainWindow::resetLayout);
+    m_viewMenu->addAction(m_resetLayoutAction);
+
+    logger.debug("Dock widgets created successfully");
+}
+
+void MainWindow::resetLayout() {
+    auto& logger = core::Logger::getInstance();
+    logger.info("Resetting dock layout to default");
+
+    // Remove all docks
+    removeDockWidget(m_navigatorDock);
+    removeDockWidget(m_propertiesDock);
+    removeDockWidget(m_logDock);
+    removeDockWidget(m_searchDock);
+    removeDockWidget(m_assistantDock);
+
+    // Re-add in default layout
+    addDockWidget(Qt::LeftDockWidgetArea, m_navigatorDock);
+    addDockWidget(Qt::RightDockWidgetArea, m_propertiesDock);
+    addDockWidget(Qt::BottomDockWidgetArea, m_logDock);
+    addDockWidget(Qt::RightDockWidgetArea, m_searchDock);
+    addDockWidget(Qt::RightDockWidgetArea, m_assistantDock);
+
+    // Tab right-side docks
+    tabifyDockWidget(m_propertiesDock, m_searchDock);
+    tabifyDockWidget(m_searchDock, m_assistantDock);
+
+    // Raise Properties tab
+    m_propertiesDock->raise();
+
+    // Show all docks
+    m_navigatorDock->show();
+    m_propertiesDock->show();
+    m_logDock->show();
+    m_searchDock->show();
+    m_assistantDock->show();
+
+    statusBar()->showMessage(tr("Layout reset to default"), 2000);
+}
+
+// Perspective save/restore
+void MainWindow::closeEvent(QCloseEvent* event) {
+    auto& logger = core::Logger::getInstance();
+    logger.debug("Saving window perspective");
+
+    QSettings settings("Bartosz W. Warzocha & Kalahari Team", "Kalahari");
+    settings.setValue("geometry", saveGeometry());
+    settings.setValue("windowState", saveState());
+
+    logger.debug("Window perspective saved");
+
+    QMainWindow::closeEvent(event);
+}
+
+void MainWindow::showEvent(QShowEvent* event) {
+    QMainWindow::showEvent(event);
+
+    if (m_firstShow) {
+        auto& logger = core::Logger::getInstance();
+        logger.debug("Restoring window perspective");
+
+        QSettings settings("Bartosz W. Warzocha & Kalahari Team", "Kalahari");
+        restoreGeometry(settings.value("geometry").toByteArray());
+        restoreState(settings.value("windowState").toByteArray());
+
+        logger.debug("Window perspective restored");
+
+        m_firstShow = false;
+    }
 }
 
 } // namespace gui